+++ plugins/wp-bizwit/resources/admin/js/app.js
@@ -1,7 +1,0 @@
-/**
- * All of the code for your admin-facing JavaScript source
- * should reside in this file.
- *
- * jQuery is available via global aliases ($ and jQuery).
- * Example: $(document).ready(function() { ... });
- */